+++ src/main/kotlin/com/zepben/auth/client/AuthProviderConfig.kt
@@ -11,7 +11,6 @@ package com.zepben.auth.client
 import com.zepben.auth.common.AuthException
 import com.zepben.auth.common.AuthMethod
 import com.zepben.auth.common.StatusCode
-import com.zepben.auth.server.TrustedIssuer
 import io.vertx.core.json.DecodeException
 import io.vertx.core.json.Json
 import io.vertx.core.json.JsonObject