+++ src/test/kotlin/com/zepben/auth/server/JWTAuthenticatorTest.kt
@@ -91,6 +91,7 @@ class JWTAuthenticatorTest {
         assertThat(authResp.message, equalTo("Unknown or untrusted issuer: https://issuer/"))
     }
 
+
     @Test
     fun `keys are updated when unknown key is provided`() {
         val jwk = Jwk("fakekid", "RSA", "RS256", "", emptyList(), "", emptyList(), "", attribs)